feat: 추억코스 화면 잠시..
